Move step menu handling to the StepMenuDelegate;
